electorator/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary>=2.8
PyJWT==1.6.4